seed refactor working
