+++ package.json
@@ -4,7 +4,7 @@
   "description": "Video game rom mover and web app",
   "main": "./dist/index.js",
   "scripts": {
-    "seed-database": "babel-node ./src/seedScripts/seedScript.js",
+    "seed-database": "babel-node ./src/fileService/seedScripts/seedScript.js",
     "run-service": "gulp backend-build&&set PORT=666&&node ./dist/backend.js",
     "run-web": "gulp build&&node ./dist/app/webApp.js"
   },
